requirements.txt
mysql-connector-python
cryptography
numpy
